Move `update_sample_with_params` to sharable location
